Use size_t for Location line/column
Change Location.line and .column from int to size_t and update callsites accordingly. Adjusted format specifiers (%d -> %zu) in checker.c, module.c, codegen.c and parser.c, switched the parser column loop index to size_t, and improved diagnostics.c to compute line number width and align the source line/caret output. These updates ensure consistent sizing/formatting for location values across the codebase.

+++ src/location.h
@@ -2,10 +2,12 @@
 #define LOCATION_H
 
 // Source location for errors
+#include <stddef.h>
+
 typedef struct {
   const char *file;
-  int line;
-  int column;
+  size_t line;
+  size_t column;
 } Location;
 
 #endif
